Create testutil.go
