+++ gateway/coprocess_test.go
@@ -141,7 +141,7 @@ func TestCoProcessMiddleware(t *testing.T) {
 
 	chain := buildCoProcessChain(spec, "hook_test", coprocess.HookType_Pre, apidef.MiddlewareDriver("python"))
 
-	session := createStandardSession()
+	session := CreateStandardSession()
 	spec.SessionManager.UpdateSession("abc", session, 60, false)
 
 	recorder := httptest.NewRecorder()
@@ -157,7 +157,7 @@ func TestCoProcessObjectPostProcess(t *testing.T) {
 
 	chain := buildCoProcessChain(spec, "hook_test_object_postprocess", coprocess.HookType_Pre, apidef.MiddlewareDriver("python"))
 
-	session := createStandardSession()
+	session := CreateStandardSession()
 	spec.SessionManager.UpdateSession("abc", session, 60, false)
 
 	recorder := httptest.NewRecorder()
@@ -211,7 +211,7 @@ func TestCoProcessAuth(t *testing.T) {
 
 	chain := buildCoProcessChain(spec, "hook_test_bad_auth", coprocess.HookType_CustomKeyCheck, apidef.MiddlewareDriver("python"))
 
-	session := createStandardSession()
+	session := CreateStandardSession()
 	spec.SessionManager.UpdateSession("abc", session, 60, false)
 
 	recorder := httptest.NewRecorder()
@@ -230,7 +230,7 @@ func TestCoProcessAuth(t *testing.T) {
 func TestCoProcessReturnOverrides(t *testing.T) {
 	spec := createSpecTest(t, basicCoProcessDef)
 	chain := buildCoProcessChain(spec, "hook_test_return_overrides", coprocess.HookType_Pre, apidef.MiddlewareDriver("python"))
-	session := createStandardSession()
+	session := CreateStandardSession()
 	spec.SessionManager.UpdateSession("abc", session, 60, false)
 
 	recorder := httptest.NewRecorder()
@@ -250,7 +250,7 @@ func TestCoProcessReturnOverrides(t *testing.T) {
 func TestCoProcessReturnOverridesErrorMessage(t *testing.T) {
 	spec := createSpecTest(t, basicCoProcessDef)
 	chain := buildCoProcessChain(spec, "hook_test_return_overrides_error", coprocess.HookType_Pre, apidef.MiddlewareDriver("python"))
-	session := createStandardSession()
+	session := CreateStandardSession()
 	spec.SessionManager.UpdateSession("abc", session, 60, false)
 
 	recorder := httptest.NewRecorder()
