+++ gateway/mw_virtual_endpoint_test.go
@@ -24,7 +24,7 @@ function testVirtData(request, session, config) {
 `
 
 func testPrepareVirtualEndpoint(js string, method string, path string, proxyOnError bool) {
-	buildAndLoadAPI(func(spec *APISpec) {
+	BuildAndLoadAPI(func(spec *APISpec) {
 		spec.Proxy.ListenPath = "/"
 
 		virtualMeta := apidef.VirtualMeta{
@@ -58,7 +58,7 @@ func testPrepareVirtualEndpoint(js string, method string, path string, proxyOnEr
 }
 
 func TestVirtualEndpoint(t *testing.T) {
-	ts := newTykTestServer()
+	ts := StartMock()
 	defer ts.Close()
 
 	testPrepareVirtualEndpoint(virtTestJS, "GET", "/virt", true)
@@ -75,7 +75,7 @@ func TestVirtualEndpoint(t *testing.T) {
 }
 
 func TestVirtualEndpoint500(t *testing.T) {
-	ts := newTykTestServer()
+	ts := StartMock()
 	defer ts.Close()
 
 	testPrepareVirtualEndpoint("abc", "GET", "/abc", false)
@@ -89,7 +89,7 @@ func TestVirtualEndpoint500(t *testing.T) {
 func BenchmarkVirtualEndpoint(b *testing.B) {
 	b.ReportAllocs()
 
-	ts := newTykTestServer()
+	ts := StartMock()
 	defer ts.Close()
 
 	testPrepareVirtualEndpoint(virtTestJS, "GET", "/virt", true)
