+++ gateway/rpc_test.go
@@ -200,7 +200,7 @@ func TestSyncAPISpecsRPCSuccess(t *testing.T) {
 	// Mock RPC
 	dispatcher := gorpc.NewDispatcher()
 	dispatcher.AddFunc("GetApiDefinitions", func(clientAddr string, dr *DefRequest) (string, error) {
-		return jsonMarshalString(buildAPI(func(spec *APISpec) {
+		return jsonMarshalString(BuildAPI(func(spec *APISpec) {
 			spec.UseKeylessAccess = false
 		})), nil
 	})
@@ -211,13 +211,13 @@ func TestSyncAPISpecsRPCSuccess(t *testing.T) {
 		return true
 	})
 	dispatcher.AddFunc("GetKey", func(clientAddr, key string) (string, error) {
-		return jsonMarshalString(createStandardSession()), nil
+		return jsonMarshalString(CreateStandardSession()), nil
 	})
 
 	t.Run("RPC is live", func(t *testing.T) {
 		rpc := startRPCMock(dispatcher)
 		defer stopRPCMock(rpc)
-		ts := newTykTestServer()
+		ts := StartMock()
 		defer ts.Close()
 
 		apiBackup, _ := LoadDefinitionsFromRPCBackup()
@@ -252,7 +252,7 @@ func TestSyncAPISpecsRPCSuccess(t *testing.T) {
 		config.SetGlobal(globalConf)
 
 		// RPC layer is down
-		ts := newTykTestServer()
+		ts := StartMock()
 		defer ts.Close()
 
 		// Wait for backup to load
@@ -279,7 +279,7 @@ func TestSyncAPISpecsRPCSuccess(t *testing.T) {
 
 		dispatcher := gorpc.NewDispatcher()
 		dispatcher.AddFunc("GetApiDefinitions", func(clientAddr string, dr *DefRequest) (string, error) {
-			return jsonMarshalString(buildAPI(
+			return jsonMarshalString(BuildAPI(
 				func(spec *APISpec) { spec.UseKeylessAccess = false },
 				func(spec *APISpec) { spec.UseKeylessAccess = false },
 			)), nil
@@ -291,12 +291,12 @@ func TestSyncAPISpecsRPCSuccess(t *testing.T) {
 			return true
 		})
 		dispatcher.AddFunc("GetKey", func(clientAddr, key string) (string, error) {
-			return jsonMarshalString(createStandardSession()), nil
+			return jsonMarshalString(CreateStandardSession()), nil
 		})
 		// Back to live
 		rpc := startRPCMock(dispatcher)
 		defer stopRPCMock(rpc)
-		ts := newTykTestServer()
+		ts := StartMock()
 		defer ts.Close()
 
 		time.Sleep(100 * time.Millisecond)
@@ -319,7 +319,7 @@ func TestSyncAPISpecsRPCSuccess(t *testing.T) {
 
 	t.Run("RPC is back, live reload", func(t *testing.T) {
 		rpc := startRPCMock(dispatcher)
-		ts := newTykTestServer()
+		ts := StartMock()
 		defer ts.Close()
 
 		time.Sleep(100 * time.Millisecond)
